Biblio Digest Redesign (#333)
* Use TranslateDate function instead of getFormattedDate to avoid having to maintain date formats in a hardcoded manner

* Function renaming and data entry

* Add literature update report card

* Add translations

* Update french translation

+++ src/utils/translate/translateService.ts
@@ -132,14 +132,14 @@ export function TranslateObject(text: string): object {
 
 /**
  * Used to translate and localize dates
- * @param dateString
+ * @param dateString with format `{YYYY-MM-DD}T{HH:MM:SS}`
  * @returns string
  */
  export function TranslateDate(dateString: string): string {
 
   const dateTimeFormat = new Intl.DateTimeFormat(language, {
     year: 'numeric',
-    month: 'long',
+    month: 'short',
     day: 'numeric',
   });
